Move interface to spi
